docker/requirements.txt
paho-mqtt
timeloop
